internship-pa/backend/requirements.txt
supabase==2.3.5
google-generativeai==0.5.4
groq==0.5.0
openai==1.14.0
together==0.2.11
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.124.0
playwright==1.43.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
pytz==2024.1
httpx==0.27.0
lxml==5.2.1
fake-useragent==1.5.1
gspread==6.1.0
